added token auth for blogs with testing

+++ controllers/blogs.js
@@ -1,30 +1,59 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const { userExtractor } = require('../utils/middleware')
+
+const jwt = require('jsonwebtoken')
+/* Now in middleware
+const getTokenFrom = request => {
+    const authorization = request.get('authorization')
+    if (authorization && authorization.startsWith('Bearer ')) {
+      return authorization.replace('Bearer ', '')
+    }
+    return null
+}*/
 
 blogRouter.get('/', async (request, response, next) => {
     console.log('Fetching all blogs')
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1 })
     response.json(blogs)
 })
 
-blogRouter.post('/', async (request, response, next) => {
+blogRouter.post('/', userExtractor, async (request, response, next) => {
     
     const body = request.body
+
+    const user = request.user
+    console.log(user)
 
     const blog = new Blog({
         title: body.title,
         author: body.author,
         url: body.url,
-        likes: body.likes || 0
+        likes: body.likes || 0,
+        user: user._id
     })
 
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
     response.json(savedBlog)
 })
 
-blogRouter.delete('/:id', async (request, response, next) => {
-    await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+blogRouter.delete('/:id', userExtractor, async (request, response, next) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    console.log(user)
+    console.log(blog)
+
+    if ( blog.user.toString() === user._id.toString() ) {
+        await Blog.findByIdAndDelete(request.params.id)
+        return response.status(204).json(blog)
+    } else {
+        return response.status(400).json({error: 'Insufficient Permissions'})
+    }
 })
 
 blogRouter.put('/:id', async (request, response, next) => {
